// node/configure/type/crawler-resource.d.ts
/* tslint:disable */
/**
 * This file was automatically generated by json-schema-to-typescript.
 * DO NOT MODIFY IT BY HAND. Instead, modify the source JSONSchema file,
 * and run json-schema-to-typescript to regenerate this file.
 */

export type NoName1 = string;
export type NoName3 = number;
export type NoName4 = string;
export type NoName5 = string;
export type NoName6 = string;

export interface NoName {
  log_dir: NoName1;
  diff: NoName2;
  view: View;
}
export interface NoName2 {
  max_line: NoName3;
  omit: NoName4;
}
export interface View {
  init: NoName5;
  log: NoName6;
}